# static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "NAAS PORT LAUNCHER",
        "description": "APIs for interacting with sandbox IAP",
        "version": "1.0.0"
    },
    "tags": [{
            "name": "Auth",
            "description": "for IAP authentication"
        },
        {
            "name": "Start a workflow",
            "description": "starts a workflow in sandbox env"
        }
    ],
    "paths": {
        "/auth": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "takes username and password then fetches authentication token",
                "parameters": [{
                        "name": "sandboxNumber",
                        "in": "query",
                        "description": "The IAP sandbox number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 17
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin@pronghorn"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated token",
                        "content": {
                            "plain/text": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "if4r98ffb7fg4904g408g0h4g="
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "plain/text": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Can't reach the IAP server"
                            }
                        }
                    }
                }
            }
        },
        "/start_wf/LNAAS_CREATE_UNI_SL_V1": {
            "post": {
                "tags": [
                    "Start a workflow"
                ],
                "summary": "takes sandbox number and payload, executes the WF and returns job-ID",
                "parameters": [{
                    "name": "sandboxNumber",
                    "in": "query",
                    "description": "The IAP sandbox number",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    },
                    "example": 17
                }],
                "requestBody": {
                    "description": "The request payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "requestPayload": {
                                        "type": "object",
                                        "allOf": [{
                                                "$ref": "./schemas/create-port-sl-v1-schema.json"
                                            },
                                            {
                                                "type": "object"
                                            }
                                        ]
                                    }
                                }
                            },
                            "examples": {
                                "example 1": {
                                    "$ref": "./examples/create_uni_sl_payload1.json"
                                },
                                "example 2": {
                                    "$ref": "./examples/create_uni_sl_payload2.json"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job ID and URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "job_id": {
                                            "type": "string"
                                        },
                                        "job_url": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "example": {
                                    "job_id": "91cb4a3272f549fe8a8732d9",
                                    "job_url": "http://titansdevcarrier.clcloud.af.qwest.net:4017/workflow_engine/viewer?job_id=91cb4a3272f549fe8a8732d9&job_status=complete&details=true"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "plain/text": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Cannot start the job"
                            }
                        }
                    }
                }
            }
        }
    }
}